requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-filter==22.1
django-graphql-jwt==0.4.0
djangorestframework==3.14.0
graphene==3.3
graphene-django==3.1.5
graphene-django-extras==1.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.8.0
whitenoise==6.6.0
